Begin GameController refactor
